requirements.txt
streamlit==0.53.0
plotly_express==0.4.1
scikit-learn==0.22.1